APPOBS-32684/Refactor + bump version
